d2com-survey/requirements.txt
# D2Com Survey System — Dependencies

# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# Auth
PyJWT==2.9.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1

# Google APIs
google-api-python-client==2.149.0
gspread==6.1.0

# Utils
python-dotenv==1.0.1
httpx==0.27.0
